+++ MVCProject/Controllers/PromotionTypeController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using MVCProject.Models;
+using Microsoft.AspNet.Identity;
 
 namespace MVCProject.Controllers
 {
@@ -17,12 +18,16 @@ namespace MVCProject.Controllers
         // GET: /PromotionType/
         public ActionResult Index()
         {
+            if (!Common.Commons.CheckLogin(Request, Response, User.Identity.GetUserName()))
+                return null;
             return View(db.PromotionTypes.ToList());
         }
 
         // GET: /PromotionType/Details/5
         public ActionResult Details(int? id)
         {
+            if (!Common.Commons.CheckLogin(Request, Response, User.Identity.GetUserName()))
+                return null;
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -38,6 +43,8 @@ namespace MVCProject.Controllers
         // GET: /PromotionType/Create
         public ActionResult Create()
         {
+            if (!Common.Commons.CheckLogin(Request, Response, User.Identity.GetUserName()))
+                return null;
             ViewBag.PromotionTypeList = Common.Commons.GetPromotionTypeList(db);
             return View();
         }
@@ -49,6 +56,8 @@ namespace MVCProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include="ProTypeName,AddType,ExRate,ID")] PromotionType promotiontype)
         {
+            if (!Common.Commons.CheckLogin(Request, Response, User.Identity.GetUserName()))
+                return null;
             if (ModelState.IsValid)
             {
                 db.PromotionTypes.Add(promotiontype);
@@ -62,6 +71,8 @@ namespace MVCProject.Controllers
         // GET: /PromotionType/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!Common.Commons.CheckLogin(Request, Response, User.Identity.GetUserName()))
+                return null;
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -81,6 +92,8 @@ namespace MVCProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include="ProTypeName,AddType,ExRate,ID")] PromotionType promotiontype)
         {
+            if (!Common.Commons.CheckLogin(Request, Response, User.Identity.GetUserName()))
+                return null;
             if (ModelState.IsValid)
             {
                 db.Entry(promotiontype).State = EntityState.Modified;
@@ -93,6 +106,8 @@ namespace MVCProject.Controllers
         // GET: /PromotionType/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!Common.Commons.CheckLogin(Request, Response, User.Identity.GetUserName()))
+                return null;
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -110,6 +125,8 @@ namespace MVCProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!Common.Commons.CheckLogin(Request, Response, User.Identity.GetUserName()))
+                return null;
             PromotionType promotiontype = db.PromotionTypes.Find(id);
             db.PromotionTypes.Remove(promotiontype);
             db.SaveChanges();
